ref: cleaning the code
