+++ app/utils/messages/success.py
@@ -1,0 +1,5 @@
+from schemas.base import BaseMessage
+
+
+def Success(detail: str) -> BaseMessage:
+    return BaseMessage(detail=detail)
